# openapi/build/generate.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Formance Stack API",
    "description": "Open, modular foundation for unique payments flows\n\n# Introduction\nThis API is documented in **OpenAPI format**.\n\n# Authentication\nFormance Stack offers one forms of authentication:\n  - OAuth2\nOAuth2 - an open protocol to allow secure authorization in a simple\nand standard method from web, mobile and desktop applications.\n<SecurityDefinitions />\n",
    "contact": {
      "name": "Formance",
      "url": "https://www.formance.com",
      "email": "support@formance.com"
    },
    "x-logo": {
      "url": "https://avatars.githubusercontent.com/u/84325077?s=200&v=4",
      "altText": "Formance"
    },
    "version": "v1.0.20230816"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "local server"
    },
    {
      "url": "https://{organization}.sandbox.formance.cloud",
      "description": "sandbox server",
      "variables": {
        "organization": {
          "description": "The organization on which the ledger is located",
          "default": ""
        }
      }
    }
  ],
  "security": [
    {
      "Authorization": []
    }
  ],
  "paths": {
...
    "/api/ledger/{ledger}/transactions": {
      "head": {
        "tags": [
          "Ledger",
          "Transactions"
        ],
        "summary": "Count the transactions from a ledger",
        "operationId": "countTransactions",
        "parameters": [
          {
            "name": "ledger",
            "in": "path",
            "description": "Name of the ledger.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ledger001"
            }
          },
          {
            "name": "reference",
            "in": "query",
            "description": "Filter transactions by reference field.",
            "schema": {
              "type": "string",
              "example": "ref:001"
            }
          },
          {
            "name": "account",
            "in": "query",
            "description": "Filter transactions with postings involving given account, either as source or destination (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter transactions with postings involving given account at source (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "description": "Filter transactions with postings involving given account at destination (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "Filter transactions that occurred after this timestamp.\nThe format is RFC3339 and is inclusive (for example, \"2023-01-02T15:04:01Z\" includes the first second of 4th minute).\n",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "Filter transactions that occurred before this timestamp.\nThe format is RFC3339 and is exclusive (for example, \"2023-01-02T15:04:01Z\" excludes the first second of 4th minute).\n",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "metadata",
            "in": "query",
            "description": "Filter transactions by metadata key value pairs. Nested objects can be used like this -> metadata[key]=value1&metadata[a.nested.key]=value2",
            "style": "deepObject",
            "schema": {
              "type": "object",
              "properties": {},
              "additionalProperties": {
                "type": "string"
              }
            },
            "example": {
              "admin": "true"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "headers": {
              "Count": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Ledger",
          "Transactions"
        ],
        "summary": "List transactions from a ledger",
        "description": "List transactions from a ledger, sorted by id in descending order.",
        "operationId": "listTransactions",
        "parameters": [
          {
            "name": "ledger",
            "in": "path",
            "description": "Name of the ledger.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ledger001"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The maximum number of results to return per page.\n",
            "example": 100,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "reference",
            "in": "query",
            "description": "Find transactions by reference field.",
            "schema": {
              "type": "string",
              "example": "ref:001"
            }
          },
          {
            "name": "account",
            "in": "query",
            "description": "Filter transactions with postings involving given account, either as source or destination (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter transactions with postings involving given account at source (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "description": "Filter transactions with postings involving given account at destination (regular expression placed between ^ and $).",
            "schema": {
              "type": "string",
              "example": "users:001"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "Filter transactions that occurred after this timestamp.\nThe format is RFC3339 and is inclusive (for example, \"2023-01-02T15:04:01Z\" includes the first second of 4th minute).\n",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "Filter transactions that occurred before this timestamp.\nThe format is RFC3339 and is exclusive (for example, \"2023-01-02T15:04:01Z\" excludes the first second of 4th minute).\n",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Parameter used in pagination requests. Maximum page size is set to 15.\nSet to the value of next for the next page of results.\nSet to the value of previous for the previous page of results.\nNo other parameters can be set when this parameter is set.\n",
            "schema": {
              "type": "string",
              "example": "aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ=="
            }
          },
          {
            "name": "metadata",
            "in": "query",
            "description": "Filter transactions by metadata key value pairs. Nested objects can be used as seen in the example below.",
            "style": "deepObject",
            "schema": {
              "type": "object",
              "properties": {},
              "additionalProperties": {
                "type": "string"
              }
            },
            "example": {
              "admin": "true"
            }
          },
          {
            "name": "expand",
            "in": "query",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsCursorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ledger",
          "Transactions"
        ],
        "summary": "Create a new transaction to a ledger",
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "ledger",
            "in": "path",
            "description": "Name of the ledger.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ledger001"
            }
          },
          {
            "name": "dryRun",
            "in": "query",
            "description": "Set the dryRun mode. dry run mode doesn't add the logs to the database or publish a message to the message broker.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Use an idempotency key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The request body must contain at least one of the following objects:\n  - `postings`: suitable for simple transactions\n  - `script`: enabling more complex transactions with Numscript\n",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTransaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTransactionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/{ledger}/transactions/{id}": {
      "get": {
        "tags": [
          "Ledger",
          "Transactions"
        ],
        "summary": "Get transaction from a ledger by its ID",
        "operationId": "getTransaction",
        "parameters": [
          {
            "name": "ledger",
            "in": "path",
            "description": "Name of the ledger.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ledger001"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Transaction ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "example": 1234
            }
          },
          {
            "name": "expand",
            "in": "query",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ledger/{ledger}/transactions/{id}/revert": {
      "post": {
        "tags": [
          "Ledger",
          "Transactions"
        ],
        "operationId": "revertTransaction",
        "summary": "Revert a ledger transaction by its ID",
        "parameters": [
          {
            "name": "ledger",
            "in": "path",
            "description": "Name of the ledger.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ledger001"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Transaction ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "example": 1234
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevertTransactionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wallets/transactions": {
      "get": {
        "operationId": "getTransactions",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "description": "The maximum number of results to return per page",
            "example": 100,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 15
            }
          },
          {
            "name": "walletID",
            "in": "query",
            "description": "A wallet ID to filter on",
            "example": "wallet1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Parameter used in pagination requests.\nSet to the value of next for the next page of results.\nSet to the value of previous for the previous page of results.\nNo other parameters can be set when the cursor is set.\n",
            "schema": {
              "type": "string",
              "example": "aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ=="
            }
          }
        ],
        "tags": [
          "Wallets"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/walletsErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransactionsCursorResponse": {
        "type": "object",
        "required": [
          "cursor"
        ],
        "properties": {
          "cursor": {
            "type": "object",
            "required": [
              "pageSize",
              "hasMore",
              "data"
            ],
            "properties": {
              "pageSize": {
                "type": "integer",
                "format": "int64",
                "minimum": 1,
                "maximum": 1000,
                "example": 15
              },
              "hasMore": {
                "type": "boolean",
                "example": false
              },
              "previous": {
                "type": "string",
                "example": "YXVsdCBhbmQgYSBtYXhpbXVtIG1heF9yZXN1bHRzLol="
              },
              "next": {
                "type": "string",
                "example": ""
              },
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ExpandedTransaction"
                }
              }
            }
          }
        }
      },
...
      "ledgerMetadata": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        },
        "example": {
          "admin": "true"
        }
      },
...
      "Posting": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "bigint",
            "minimum": 0,
            "example": 100
          },
          "asset": {
            "type": "string",
            "example": "COIN"
          },
          "destination": {
            "type": "string",
            "example": "users:002"
          },
          "source": {
            "type": "string",
            "example": "users:001"
          }
        },
        "required": [
          "amount",
          "asset",
          "destination",
          "source"
        ]
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Posting"
            }
          },
          "reference": {
            "type": "string",
            "example": "ref:001"
          },
          "metadata": {
            "$ref": "#/components/schemas/ledgerMetadata"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "reverted": {
            "type": "boolean"
          }
        },
        "required": [
          "postings",
          "timestamp",
          "id",
          "metadata",
          "reverted"
        ]
      },
      "ExpandedTransaction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Transaction"
          },
          {
            "type": "object",
            "properties": {
              "preCommitVolumes": {
                "$ref": "#/components/schemas/AggregatedVolumes"
              },
              "postCommitVolumes": {
                "$ref": "#/components/schemas/AggregatedVolumes"
              }
            }
          }
        ]
      },
      "PostTransaction": {
        "type": "object",
        "required": [
          "metadata"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Posting"
            }
          },
          "script": {
            "type": "object",
            "properties": {
              "plain": {
                "type": "string",
                "example": "vars {\naccount $user\n}\nsend [COIN 10] (\n\tsource = @world\n\tdestination = $user\n)\n"
              },
              "vars": {
                "type": "object",
                "properties": {},
                "example": {
                  "user": "users:042"
                }
              }
            },
            "required": [
              "plain"
            ]
          },
          "reference": {
            "type": "string",
            "example": "ref:001"
          },
          "metadata": {
            "$ref": "#/components/schemas/ledgerMetadata"
          }
        }
      },
...
      "CreateTransactionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Transaction"
          }
        },
        "type": "object",
        "required": [
          "data"
        ]
      },
      "RevertTransactionResponse": {
        "$ref": "#/components/schemas/CreateTransactionResponse"
      },
      "GetTransactionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExpandedTransaction"
          }
        },
        "type": "object",
        "required": [
          "data"
        ]
      },
...
      "Volumes": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Volume"
        },
        "example": {
          "USD": {
            "input": 100,
            "output": 10,
            "balance": 90
          },
          "EUR": {
            "input": 100,
            "output": 10,
            "balance": 90
          }
        }
      },
      "AggregatedVolumes": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Volumes"
        },
        "example": {
          "orders:1": {
            "USD": {
              "input": 100,
              "output": 10,
              "balance": 90
            }
          },
          "orders:2": {
            "USD": {
              "input": 100,
              "output": 10,
              "balance": 90
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "errorCode",
          "errorMessage"
        ],
        "properties": {
          "errorCode": {
            "$ref": "#/components/schemas/ErrorsEnum"
          },
          "errorMessage": {
            "type": "string",
            "example": "[VALIDATION] invalid 'cursor' query param"
          },
          "details": {
            "type": "string",
            "example": "https://play.numscript.org/?payload=eyJlcnJvciI6ImFjY291bnQgaGFkIGluc3VmZmljaWVudCBmdW5kcyJ9"
          }
        }
      },
      "ErrorsEnum": {
        "type": "string",
        "enum": [
          "INTERNAL",
          "INSUFFICIENT_FUND",
          "VALIDATION",
          "CONFLICT",
          "NO_SCRIPT",
          "COMPILATION_FAILED",
          "METADATA_OVERRIDE",
          "NOT_FOUND",
          "CONTEXT_CANCELLED",
          "STORE"
        ],
        "example": "VALIDATION"
      },
...
      "walletsAggregatedVolumes": {
        "type": "object",
        "x-go-type": {
          "type": "AggregatedVolumes"
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/walletsVolumes"
        }
      },
      "walletsTransaction": {
        "type": "object",
        "properties": {
          "ledger": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Posting"
            }
          },
          "reference": {
            "type": "string",
            "example": "ref:001"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Metadata associated with the wallet."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "preCommitVolumes": {
            "$ref": "#/components/schemas/walletsAggregatedVolumes"
          },
          "postCommitVolumes": {
            "$ref": "#/components/schemas/walletsAggregatedVolumes"
          }
        },
        "required": [
          "postings",
          "timestamp",
          "ix",
          "metadata"
        ]
      },
      "walletsCursor": {
        "type": "object",
        "required": [
          "pageSize"
        ],
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 1000,
            "example": 15
          },
          "hasMore": {
            "type": "boolean",
            "example": false
          },
          "previous": {
            "type": "string",
            "example": "YXVsdCBhbmQgYSBtYXhpbXVtIG1heF9yZXN1bHRzLol="
          },
          "next": {
            "type": "string",
            "example": ""
          }
        }
      },
      "GetTransactionsResponse": {
        "type": "object",
        "required": [
          "cursor"
        ],
        "properties": {
          "cursor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/walletsCursor"
              },
              {
                "properties": {
                  "data": {
                    "items": {
                      "$ref": "#/components/schemas/walletsTransaction"
                    },
                    "type": "array"
                  }
                },
                "type": "object",
                "required": [
                  "data"
                ]
              }
            ]
          }
        }
      },
...
      "walletsErrorResponse": {
        "type": "object",
        "required": [
          "errorCode",
          "errorMessage"
        ],
        "properties": {
          "errorCode": {
            "type": "string",
            "enum": [
              "VALIDATION",
              "INTERNAL_ERROR",
              "INSUFFICIENT_FUND",
              "HOLD_CLOSED"
            ]
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "",
            "refreshUrl": "",
            "scopes": {}
          }
        }
      }
    },
...
  },
...
}